requirements.txt
discord.py
webcolors